added ability for converting methods for types from "System" and "NUnit.Framework" namespaces
