+++ PythonExamplesPorter/PythonExamplesPorterApp/Converter/ExpressionConverter.cs
@@ -4,7 +4,6 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using PythonExamplesPorterApp.Checker;
 using PythonExamplesPorterApp.Common;
-using PythonExamplesPorterApp.Config;
 
 namespace PythonExamplesPorterApp.Converter
 {
@@ -57,6 +56,7 @@ namespace PythonExamplesPorterApp.Converter
                     break;
                 default:
                     _buffer.Append("<<<some expression>>>");
+                    //throw new UnsupportedSyntaxException($"Unsupported expression: {expression.Kind()}");
                     break;
             }
         }
@@ -71,7 +71,11 @@ namespace PythonExamplesPorterApp.Converter
             OperationResult<TypeResolveData> resolveResult = _externalEntityResolver.ResolveType(type);
             if (!resolveResult.Success)
                 throw new UnsupportedSyntaxException(resolveResult.Reason);
-            ProcessTypeResolveData(resolveResult.Data!);
+            TypeResolveData data = resolveResult.Data!;
+            // TODO (std_string) : i return empty type for system type because we need in additional analysis - think about approach
+            if (String.IsNullOrEmpty(data.TypeName))
+                throw new UnsupportedSyntaxException($"Unsupported type: {type}");
+            ProcessTypeResolveData(data);
             ExpressionConverter expressionConverter = new ExpressionConverter(_model, _appData);
             String[] arguments = ConvertArgumentList(expressionConverter, argumentList);
             _buffer.Append($"({String.Join(", ", arguments)})");
@@ -167,98 +171,5 @@ namespace PythonExamplesPorterApp.Converter
         private readonly ExternalEntityResolver _externalEntityResolver;
         private readonly StringBuilder _buffer;
         private readonly IDictionary<String, String> _importData;
-    }
-
-    internal record TypeResolveData(String TypeName, String ModuleName);
-
-    internal record MethodCallResolveData(String Call, String ModuleName);
-
-    // TODO (std_string) : we must implement this functionality via Strategy pattern
-    internal class ExternalEntityResolver
-    {
-        public ExternalEntityResolver(SemanticModel model, AppData appData)
-        {
-            _model = model;
-            _appData = appData;
-        }
-
-        public OperationResult<TypeResolveData> ResolveType(TypeSyntax type)
-        {
-            SymbolInfo symbolInfo = _model.GetSymbolInfo(type);
-            ISymbol? typeInfo = symbolInfo.Symbol;
-            if (typeInfo == null)
-                return new OperationResult<TypeResolveData>(false, $"Unrecognizable type: {type}");
-            return typeInfo switch
-            {
-                INamedTypeSymbol typeSymbol => ResolveType(typeSymbol),
-                _ => new OperationResult<TypeResolveData>(false, $"Unsupported {typeInfo.Kind} kind of symbol info for {type}")
-            };
-        }
-
-        public OperationResult<TypeResolveData> ResolveType(INamedTypeSymbol typeSymbol)
-        {
-            // TODO (std_string) : think about check containing assemblies
-            String[] knownNamespaces = _appData.AppConfig.GetSourceDetails().KnownNamespaces ?? Array.Empty<String>();
-            String sourceNamespaceName = typeSymbol.ContainingNamespace.ToDisplayString();
-            String sourceTypeName = typeSymbol.Name;
-            Boolean isSupportedType = knownNamespaces.Any(sourceNamespaceName.StartsWith);
-            if (!isSupportedType)
-                return new OperationResult<TypeResolveData>(false, $"Unsupported type: {sourceNamespaceName}.{sourceTypeName}");
-            String destModuleName = NameTransformer.TransformNamespaceName(sourceNamespaceName);
-            String destTypeName = NameTransformer.TransformClassName(sourceTypeName);
-            return new OperationResult<TypeResolveData>(true, "", new TypeResolveData(destTypeName, destModuleName));
-        }
-
-        public OperationResult<MethodCallResolveData> ResolveMethodCall(ExpressionSyntax target, String targetRepresentation, SimpleNameSyntax name, String[] arguments)
-        {
-            // TODO (std_string) : think about check containing assemblies
-            String[] knownNamespaces = _appData.AppConfig.GetSourceDetails().KnownNamespaces ?? Array.Empty<String>();
-            OperationResult<SourceType> targetTypeResult = ExtractMethodTargetType(target);
-            if (!targetTypeResult.Success)
-                return new OperationResult<MethodCallResolveData>(false, targetTypeResult.Reason);
-            SourceType sourceType = targetTypeResult.Data!;
-            String typeFullName = $"{sourceType.NamespaceName}.{sourceType.TypeName}";
-            Boolean isSupportedType = knownNamespaces.Any(sourceType.NamespaceName.StartsWith);
-            if (!isSupportedType)
-                return new OperationResult<MethodCallResolveData>(false, $"Unsupported type: {typeFullName}");
-            SymbolInfo nameInfo = _model.GetSymbolInfo(name);
-            switch (nameInfo.Symbol)
-            {
-                case null:
-                    return new OperationResult<MethodCallResolveData>(false, $"Unrecognizable method \"{name.Identifier}\" for type \"{typeFullName}\"");
-                case IMethodSymbol methodSymbol:
-                    String methodName = NameTransformer.TransformMethodName(methodSymbol.Name);
-                    String args = String.Join(", ", arguments);
-                    String methodCall = String.Concat(targetRepresentation, ".", methodName, "(", args, ")");
-                    MethodCallResolveData resolveData = new MethodCallResolveData(methodCall, "");
-                    return new OperationResult<MethodCallResolveData>(true, "", resolveData);
-                default:
-                    return new OperationResult<MethodCallResolveData>(false, $"Unsupported method \"{typeFullName}\"");
-            }
-        }
-
-        private record SourceType(String NamespaceName, String TypeName);
-
-        private OperationResult<SourceType> ExtractMethodTargetType(ExpressionSyntax target)
-        {
-            return ExtractMethodTargetType(target, _model.GetSymbolInfo(target).Symbol);
-        }
-
-        private OperationResult<SourceType> ExtractMethodTargetType(ExpressionSyntax target, ISymbol? targetSymbol)
-        {
-            return targetSymbol switch
-            {
-                null =>
-                    new OperationResult<SourceType>(false, $"Unrecognizable method target type for \"{target}\""),
-                ILocalSymbol localSymbol => ExtractMethodTargetType(target, localSymbol.Type),
-                INamedTypeSymbol typeSymbol =>
-                    new OperationResult<SourceType>(true, "", new SourceType(typeSymbol.ContainingNamespace.ToDisplayString(), typeSymbol.Name)),
-                _ =>
-                    new OperationResult<SourceType>(false, $"Unsupported method target type for \"{target}\"")
-            };
-        }
-
-        private readonly SemanticModel _model;
-        private readonly AppData _appData;
     }
 }
